#6: remove need for separate probability
